initial integration / page tests
